Add files via upload
This repository contains a straightforward yet entertaining Stone-Paper-Scissors game. It is designed with HTML, CSS, and JavaScript and lets you compete against the machine in a game of wits and unpredictability. Determine the winner, pick your weapon as me, and view the computer's selection as you! Excellent for novices wishing to practice the principles of web construction or anyone who wants to play a quick, nostalgic game.
All suggestions for improvement of code are welcome.